wip: Move Merge Models to new panel & use new model fetch
